Honor script time limit when calling shutdown functions
A time limit can be set on PHP script execution via set_time_limit() (or .ini file).
When the time limit is reached, the OS will notify PHP and a 'timed_out' flag is set.
While this flag is regularly checked when executing PHP code, once the end of the script
is reached, it is not checked while invoking shutdown functions (registered via
register_shutdown_function).

Of course, if the shutdown functions are implemented *in* PHP, then the script time limit
will be checked while they are running and will take effect. But if the shutdown functions
are built-in (implemented in C), it will not.

Since the shutdown functions are invoked through zend_call_function, add a check of the
'timed_out' flag there. Then, the script time limit will be respected when *entering*
each shutdown function. The fact still remains that if a shutdown function is built-in and
runs for a long time, script execution will not time out until it finishes and the
interpreter tries to invoke the next one.

Still, the behavior of scripts with execution time limits will be more consistent after
this patch. To make the execution time-out feature work even more precisely, it would
be necessary to scrutinize all the built-in functions and add checks of the 'timed_out'
flag in any which can run for a long time. That might not be worth the effort, though.