Simplified handling of hard timeouts
Previously, the same timeout handler function was used both for the initial
(or "soft") script execution timeout and the "hard" timeout if the script is taking
too long to run shutdown functions, etc. after the initial timeout. This meant that
when the timeout handler was invoked, it needed to detect whether another timeout
had already occurred previously.

This is more straightforward; there is one handler for soft timeouts and one for
hard timeouts. The handler for soft timeout *schedules* the one for hard timeout.