Don't use Zend signal handling to protect shared memory in ZendAccelerator.c
Zend signal handling was added in PHP 5.4 to protect against signal handlers running
at inopportune times and causing bugs. The details are in this RFC:

https://wiki.php.net/rfc/zendsignals

In short, the handlers for 7 signals of concern are saved and replaced with a generic
handler which delegates to the specific handlers. Inside 'critical sections', however,
the generic handler puts all information regarding a signal on a queue and just returns.
At the end of the critical section, all pending signals on the queue are processed and
the specific handlers are called.

However, in PHP 7.1, the `vm_interrupt` flag was added which also protects against
script execution timeouts, etc. occurring at wrong times. This eliminated most of the
use cases of Zend signal handling. The one which has remained until now is accessing
shared memory in OPCache. If that can be eliminated, there will be no need for Zend
signal handling any more and a subsystem can be unceremoniously ripped out.

The funny thing about the whole idea of Zend signal handling is... it seems to
duplicate what Unix kernels already do. Each process/thread in Unix already has a
signal mask which can be used to block signals from being delivered at inopportune
times. If a signal arrives when it is masked, the kernel will store it and only
deliver it once it is unmasked. So rather than going through the whole dance of
storing signals on a queue and unqueueing them later... let the kernel do its job.