Remove useless struct: mbfl_filt_tl_jisx0201_jisx0208_param
This struct contained an unused 'next_filter' field. The only field which was used
was the integral 'mode'. Since only one field in the struct was used, it is better
to inline it and just store the integer directly where it is needed.

Further, while an instance of this struct was included in the struct
mbfl_filt_conv_wchar_cp50220_ctx, it was never actually used there. It was just
taking up memory for nothing.