Fix buggy support for negative 'start' argument to mb_strimwidth
According to the document, strimwidth accepts a negative value for the 'start'
parameter, which means it should count backwards the given number of characters
from the end of the input string and use that as its starting position.

This is, frankly, a bizarre feature. I am struggling to think of when it would
be useful for this specific function. But since it exists, presumably it should
*work*, right? Actually, it has been buggy since it was first implemented in
2016.

Negative 'start' offsets were implemented with this code:

    if (from < 0) {
      from += mbfl_strwidth(&string);
    }

Uh oh. That is clearly intended to convert a negative character offset to an
equivalent positive offset (counting forward from the beginning, rather than
backwards from the end). But it does it by adding the string *width* and not
the string *length*, which are different numbers if there are any fullwidth
East Asian characters in the string.

The result is that for each fullwidth character in the string, the
starting position is 'bumped forward' by one character from where it should
actually be if negative starting offsets were implemented correctly.

Unfortunately, the unit tests for this feature are also wrong. Examining
the inputs and expected outputs clearly shows that they do not agree with
the documentation. I can only conclude that the author created these
tests by running the function on some input text and taking whatever
it produced as the "correct" answer. "Testing" code in this way is
essentially no different from just declaring that it is right, "because
I said so."

This might have been noticed earlier, if not for another bug which I
will address in the next commit.

Side note: I found that my text editor mangled the EUC-JP text in the
test case, and converted it to a hex string instead.