Return values of 'filter flush' functions are never used
After running some text through a 'filter', the last step is to call its
'flush function'. This is done because some filters cache one or more
characters rather than pushing them to the output right away. (That is
because the encoding of some characters may depend on what comes before
or after.)

The flush functions all returned an integral status code... but nothing
actually *checks* the return values. When running through a buffer of
input data and pushing it into a filter, the status codes are typically
used to fail fast if the input data is invalid. But by the time we
flush the filter, all the input has already been consumed, so there
is no possibility of failing fast.

So... simplify things by making flush functions return void.