Canonicalize flags received by mb_convert_kana
mb_convert_kana receives flags which specify what conversions it should perform.
Some of these flags subsume others; for example, 'A' does everything which 'R' and
'N' do, and more.

In cases where redundant flags are provided (like "AR" instead of just "A"),
canonicalize them to the smallest set which has the same effect.

In some cases, this will reduce the number of tests which need to be evaluated
for each character when actually performing the conversion.