Slight simplification to wchar -> SJIS conversion
Some conditions were evaluated redundantly; for example, if any of the `if` conditions
above the line `if (c == 0)` were true, then that condition will certainly be false.
So it is better written as `else if (c == 0)`.

Instead of setting `s1` to a negative number and later testing if it is negative or
not, we can just take the needed action at the points where it could possibly be
set to a negative value.