Offset and mask values in mb_encode_numericentity convmap do not affect which ranges apply
This is a subtle change in behavior for mbstring, but a good one. First, some background:

mb_encode_numericentity takes a series of integers called a 'convmap'. Each group of
four integers specifies a range of codepoints which should be encoded as HTML numeric
entities, and further specifies some transformations which should be performed on them.

In particular, the first two integers in each group are low and high codepoint values
to encode. The third is a (signed) offset to add to a codepoint before encoding it,
and the fourth is a bitmask to apply to the codepoint before encoding.

After applying the offset and bitmask, legacy mbstring checked that the resulting value
was non-negative before encoding it. This test is hereby removed, and here's the reasons:

1. The documentation does not indicate in any way that the offset and mask should affect
   which codepoints are actually converted. Rather, it implies the opposite.

2. The argument names do not imply in any way that they should have such an effect.

3. Common sense does not imply that they should have such an effect.

4. There is no apparent use case for this behavior.

5. All existing tests pass with the 'non-negative' condition removed.

6. Even if we fly in the face of good sense and stubbornly say that the legacy behavior
   should remain... the legacy behavior was broken. It created a strange dependency on
   the bit width of a C `int` on the target platform. Technically, it also creates a
   dependency on the representation of negative integers, since the C standard allows
   three possible representations... but all platforms of interest use two's-complement.
   Anyways, the dependency on the bit width of `int` is clearly undesirable.

Just say NO to bitwise ops on signed integers!